model_src/Model/requirements.txt
anthropic>=0.18.0
playwright>=1.40.0
pillow>=10.0.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
feedparser>=6.0.0
requests>=2.31.0
